[MR] From front into develop

See merge request s03-final/s03p31a303!277

Former-commit-id: 5cdc0d36aaae68baac853bd7aadccee45d939784 [formerly 04989a2fef646156b52bd9a3bdb5bc373ac94dbf] [formerly a8b35aed724cd1f2bc56a3ad48c71360be4e8d6a [formerly b92fdb63efd3821bf5883143441ecadbdfb310f4]]
Former-commit-id: d624339dd5d810ff8cb3714d4d3b42c558ee5175 [formerly 38228466ccb286ffcea0db86c1218d321653ae3e]
Former-commit-id: f4c12fb8be0da40850e0acad07422abef21b2a9c

+++ frontend/RunWithMe/src/views/app/mypages/mypageTab.vue
@@ -175,18 +175,29 @@ export default {
   methods: {
     ...mapMutations(["closeSidebar"]),
     deleteRunning(deleteRun) {
-      http.delete(`/runnings/${deleteRun.runningId}`)
-      .then(data =>{
-        Swal.fire({
-          icon:'success',
-          text:'런닝이 삭제되었습니다.',
-          showConfirmButton:false,
-          timer:500,
-        })
-        this.getRunningsbyArea()
-        this.getRunnings()
+      Swal.fire({
+        title: '삭제 확인',
+        text: "해당 런닝은 복구할수 없습니다",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: '삭제'
+      }).then((result) => {
+        if (result.isConfirmed) {
+           http.delete(`/runnings/${deleteRun.runningId}`)
+          .then(data =>{
+            Swal.fire({
+              icon:'success',
+              text:'런닝이 삭제되었습니다.',
+              showConfirmButton:false,
+              timer:500,
+            })
+            this.getRunningsbyArea()
+            this.getRunnings()
+          })  
+        }
       })
-
     },
     getRunningsbyArea(){
       http.get(`runnings/areas`)
